server: Rewrite in TypeScript
Oh boy.

The changes introduced in this commit are _way_ too numerous to
summarise adequately.

The general reasoning behind converting to TypeScript goes something
like this: The server contains _all_ the "real" business logic. It's
what interacts with the storage, after all, and is ultimately
responsible for ensuring the correctness and integrity of user data.
While JS permits extremely rapid development and prototyping, it also
hides crucial typing bugs. In the process of rewriting the server, I
discovered several bugs related to type misuse - most commonly,
forgetting that GraphQL's `ID` type is a string, not a number.

The structure and organisation of the server will change as I become
more familiar and comfortable with TypeScript. This commit is close to a
1-to-1 port of the server, though the validator logic has been reworked.