server: Validate lemmas outside LemmaMut and parallelise more things
This definitely should have been two separate commits.

The basic issue with `LemmaMut.ensureExists` was that it performed
validation of the term - or rather, that "validation" also involves
potentially changing the value. This means `LemmaMut` can alter the term
into a slightly different value, without communicating this back to the
definition, which then uses the original input value. Likewise when
inserting the term of a derived definition.

Instead, perform validation of terms outside `LemmaMut`. The new type
`ValidTerm` contains a validated term, as a means of leveraging the type
system to ensure we pass validated terms into `LemmaMut`.

In addition, the new method `ensureAllExist` can be used to create
multiple lemmas in one go. This is useful chiefly for derived
definitions, which otherwise need up to three queries apiece (one to
check for existing lemma, one to insert a new, and one to recalculate
the language's lemma count). The new method drastically cuts down on the
number of queries required, making definition creation significantly
faster.