server: Add `getRequired` to DB adaptor and use in code
A fair number of queries, such as those that use aggregate functions,
are guaranteed to return at least one row. For these, it's tedious to
perform checking of the return value of `get`, which can be null, and
unsafe to assume it'll be fine by casting. The new method abstracts
verification away.