server: Make 'onDelete' action required
I found it confusing to read the database schema. Even though I know
'restrict' is the default, I kept forgetting and had to think a bit
extra about it. So making `onDelete` required seems like a reasonable
choice; that way, there is never any ambiguity.

In addition, I discovered SQLite defaults to `NO ACTION`, which really
does mean "do nothing" and leaves the database in an inconsistent state.
The checks in the code _should_ ensure no impossible deletes are even
attempted, but the whole point of `ON DELETE RESTRICT` is that the data
layer can enforce correctness too.