server: Add tags to definitions
With this commit, tags can be attached to individual definitions. They
cannot be attached to derived definitions. The `Language.tags` field
allows you to discover all tags used by all definitions in the language,
as opposed to `Query.tags` which lists all tags in the dictionary.

This commit does *not* introduce filtering by tag. That will come at a
later stage, when I have a better idea of what kind of filtering I
actually need and how to manage full-text search.