admin-ui/menu: Render phantom of activated item
When a menu item is activated by any means, render an identical copy of
it that fades out quickly. This provides a nice bit of visual feedback
to the user that the item was indeed selected. This feedback is useful
particularly when navigating by keyboard, as the item's active state has
no ability to activate before the menu closes.