ipa: Skip 'key' prop when tree has 'path'
It is not actually necessary to have a `key` when there is a `path`, as
`key === path[0]`. This saves a small amount of space in the final JSON
file.

> Why not use `key` or `path` everywhere, since they now do the same thing?

I will experiment with combining siblings that have identical subtrees,
to reduce repetition of data. Essentially I want to make it possible to
end up with a tree like:

```
        .-> s -.
velari -|      |-> ed
        '-> z -'
```

Whether this is useful or not, I don't yet know.