ipa: Fix confusing use of terminology
The word 'term' was used to refer both to searchable terms associated
with an IPA character as well as to words in the search query. This
commit normalises things so that query words are referred to as 'query
word' or sometimes just 'query' for short, and reserves 'term' for
searchable terms.