server: Save historical table layout versions when needed
Okay, so this _definitely_ should have been two separate things
entirely.

The first major change is as the title suggests. Historical versions of
inflection table layouts are now stored, and many of the SQL tables were
changed to accomodate this. The previous layout of an inflection table
is retained if there is at least one definition that uses it. The
reasons for doing this are described in the GraphQL schema, but to
summarise:

1. Inflected forms may be added as derived definitions, and
   recalculating new forms for every definition that uses a table is
   potentially quite costly.
2. It is exceedingly difficult for the user to predict the effect of
   altering an inflection table across the entire dictionary, especially
   if the table is used by many definitions.

Previously, editing of tables was simply prohibited while the table was
in use, which I find less flexible than I would like Condict to be.

Historical table layout versions are deleted as soon as they are no
longer needed. Definitions that use an older layout version can be
upgraded to the current version in the `editDefinition` mutation, by
setting `upgradeTableLayout: true` for the table.

---

In making this change, I found it very confusing how every ID everywhere
was either a `number` or a `string`, yet there is clearly a difference
between the ID of a lemma and that of a definition. The new `IdOf<T>`
and `InputIdOf<T>` can be used with number and string values,
respectively, to represent IDs for specific purposes. Different ID types
are not compatible, so attempting to pass an `IdOf<'InflectionTable'>`
into an `IdOf<'InflectionTableLayout'>` now causes a compile-time error.
This even helped me fix an existing bug!

---

The way types, schemas, models, mutators and the rest are structured
will very likely change in the future, as it's a bit messy currently.