ui: Allow Menu to create its own MenuManager
This drastically simplifies implementation of menus when you're not
using MenuTrigger or ContextMenuTrigger. Without this, you have to take
so many refs to so many different things and it just becomes messy. The
`onClose` prop on the menu can be used to detect when the menu is
closed, negating the need for a MenuManager entirely.

The only downside is that the Menu component now mounts several extra
components, but that should make very little difference in practice.