server: Use generated GraphQL types
To make the schema available to the app, whenever I start working on
that, move it into ./graphql-schema in the repository root. Since all
the GraphQL-based types are now in their own file, a very large amount
of code had to be updated to refer to those types. In addition, this
allowed me to remove quite a bit of now-useless duplicated definitions.

All uses of `ID` in GraphQL have been replaced with custom scalar types.
These are implemented using the `@id` directive, which takes care of
adding parsing and serialisation behaviour to the scalar types.

Since GraphQL files are now in a totally separate directory, resolvers
don't need to have their own folders under `schema/`, so move them into
their own `resolvers/`.

And finally, the entire GraphQL schema has moved into the more
descriptively named `graphql` directory, instead of the rather vague
`schema` (which was especially confusing in the presence of
`database/schema`).