graphql-typer: Add ability to generate client-side types
The README should hopefully explain what this change aims to accomplish.
A caveat: while almost all operation types are supported, it is still
not possible to use `@skip` or `@include` with a non-constant condition.
It's easy enough to make some fields optional in some situations, but
when you then get to fragments that are conditional and some of the
fields selected by the fragment may be included unconditionally by
another selection set... it gets messy.

Also this code is a little bit of a mess, not least because of all the
text formatting nonsense everywhere. Next step is to write a slightly
more abstract TypeScript type generator, so I can create appropriate
structures for that instead of constantly appending to a TextBuilder in
a non-ideal manner.