server: Move Apollo server stuff out of CondictServer
There's now a separate CondictHttpServer that creates the ApolloServer
and takes care of queries and such. The CondictServer class now holds
the schema and creates context values, allowing it to be reused with any
number of implementations (one could even use GraphQL's `execute` on the
schema directly, if one desired).

The `export`, `import` and `getTableSchema` functions will eventually be
moved out of CondictServer as well, to become generic utility functions.
Keeps the server clean.