graphql-typer: Separate types for each kind of operation
Instead of a single generic `Query` type, generate one type per defined
operation kind. This way you can target different kinds of operations in
functions that take them, e.g.:

  executeQuery<Q extends Query<any, any>>(
    query: Q,
    args: OperationArgs<Q>
  ): OperationResult<Q>

for a function that only accepts queries (not mutations or
subscriptions), or this:

  subscribe<S extends Subscription<any, any>>(
    subscription: S,
    args: OperationArgs<S>
  ): Promise<AsyncIterator<OperationResult<S>>

for a function that only deals with subscriptions.

The generated type names correspond to those used in the GraphQL schema.
In other words, if you define a schema like

  schema {
    query: RootQueryFloop
  }

then the generated TypeScript type for queries will be anmed
`RootQueryFloop`.