server: New pagination helper function
This commit also moves the Connection type to the model folder, where it
belongs; removes the `paginate` method from Adaptor, where it clearly
didn't belong; and updates every model to make use of the new function.

This new helper function takes an optional `GraphQLResolverInfo`, which
is used to skip fields that will never be needed. If you never query
`nodes`, there is no need to even attempt to fetch them. Note that it
does *not* take `@skip` or `@include` into account (it always includes
those fields if they are mentioned).