server: Remove logger options from ServerConfig
Since `CondictServer` accepts a `Logger` as a constructor argument, it
makes little sense to require the server config to also specifiy logger
options. I mean, it doesn't actually make use of those options in any
way. Logger options only matter when running Condict from the CLI.

As a result, we now have a separate `ServerConfigWithLogger` along with
a new function for validating one of those specifically.