ui: Add support for `force-focus` class on various components
Sometimes you want to force focus to be visible even when it's actually
been moved elsewhere, e.g. when a context menu is open. The
`force-focus` class is treated the same as `:focus`, and can be set by
code that requires the UI element to appear focused even if it isn't.

Obviously this should be used sparingly.

It is not possible to force the selected text inside a textbox to stay
highlighted.