ui: Use subpages for each component demo
The page was getting quite long and hard to navigate, and it's very easy
to lose track of where you were when it hot-reloads. This new layout
needs a lot of work to look more presentable, but it _is_ only a dev
demo, so I guess it's okay for now.