ui: Roll our own focus management components
The previously used `focus-trap` library has certain limitations that do
not play well with Condict. The notion of "make everything except this
focusable" works fine in some situations (that's why the new components
include a `<FocusTrap>`), but sometimes you also want a "make this thing
NOT focusable", which is what `<FocusScope>` solves. Having multiple
competing focus managers on the page results in needless complexity, so
I decided to implement my own from scratch.

Dependent code has no been updated. In particular the menu system needs
to be changed to work with the new focus manager.

There are many differences between the previous focus trap and mine.
Describing them all is beyond the scope of this commit.