ui: Return ref from focus manager instead of taking it
There was a risk that the ref could be assigned a new element without us
noticing. For example, if the provided ref was passed to a ref-forwarded
component that conditionally renders one thing or another, we would be
unable to react to such a change unless the `FocusScope` or `FocusTrap`
happens to be updated in such a way that it triggers a scope change. By
providing our own ref, we can be notified whenever an element is
assigned to it, and update the scopes accordingly.