ui: Some elements cannot be `:enabled`
Some elements can never be `:enabled`, such as anchors, audio and video,
meaning the `:enabled` selector never matches. Obviously we still want
these to be reachable, so don't use `:enabled` on them. Interestingly,
`:not(:disabled)` is *not* equivalent to `:enabled`, so we can use that
in cases where we don't know the tag name, like with `[tabindex]`.