ui: Restructure Checkbox, Radio and Switch
The new structure allows us to rely on `:checked`, `:indeterminate` and
`:disabled` to communicate the control's state. The resulting styles are
a bit more complex, but more accurate. For Checkbox and Radio, this also
means they can now be used as uncontrolled components, like the native
`<input>`.

Additionally, focus styles are applied directly to the marker (the
circle/square/switch before the label), resulting in a somewhat more
pleasant appearance.