table-editor: Improve derived cell names
Data cells that span multiple rows or columns now get their names from
header cells that completely overlap the data cell's rows or columns, as
opposed to only looking in the "origin" row/column. For instance, in
this table:

  +----------+----------+------------+
  |          | Ergative | Absolutive |
  +----------+----------+------------+
  | Singular | A        |            |
  +----------+----------+ C          |
  | Plural   | B        |            |
  +----------+----------+------------+

the forms A and B will be named as before ("Singular ergative" and
"Plural ergative", respectively), while C will be named simply
"Absolutive", and not "Singular absolutive" as before.