table-editor: Rewrite from the ground up
The motivation behind this rewrite is to get rid of our dependency on
`immutable`. The library hasn't been updated for over two years and
appears to be abandoned. By contrast, `immer` is in active development,
much more stable, and slightly faster to boot. My personal opinion is
that the code ends up much cleaner and easier to follow.

This rewrite also reorganises internals, to get rid of the horrible
hacky `make*` functions that generated components. We can do it much
better with generics and React context. Some things that should be
abstracted and shared are now abstracted and shared, too, reducing code
repetition, of which there was quite a bit before.