server: Change formatted text format
Formatted text, as seen in definition descriptions and the like, was
stored in a format that was not ideal. Working with text offsets to
format text requires complicated Unicode operations and lots of string
slicing, and it's very difficult to ensure all inlines are correctly
nested and the like. Additionally, it maps poorly onto the chosen rich
text editor (Slate), requiring conversion between the formats.

The new format is much closer to that used by Slate, and has the benefit
of being easier to work with too. It does require somewhat more complex
queries that return more data for unformatted text; this may be improved
in the future.