server: Remove support for MySQL
The only permitted storage backend is now SQLite.

There are a few reasons for this change.

Firstly, full-text search. SQLite is the only database that has
perfectly reasonable full-text search functionality out of the box.
MySQL's implementation would require significant configuration to be
usable by Condict, and most of that work would fall on the database
owner to take care of. Some of the changes required may not even be
possible to make without ruining other things running on the same
database. In contrast, SQLite's implementation is fully self-contained,
and sufficiently customisable as far as tokenisation and Unicode are
concerned. It's not *that* sophisticated, but it's good enough. In
addition, SQLite has such groundbreaking features as custom column
weights - very useful when you want e.g. the title to be more important
than the description, but you definitely still want the entire document
to be searchable in one go. To get SQLite's featuers when MySQL is the
backing storage, we'd basically need a separate SQLite database
alongside, at which point it's meaningless to even split them up.

Secondly, technical debt. It is costly to maintain two separate database
implementations, with schema generation and everything. Trying to design
a schema that is simultaneously compatible with both *and* optimised for
the strengths of all drivers is difficult. MySQL has more restrictive
text column lengths unless you (possibly inadvisedly) use `MEDIUMTEXT`
or `LONGTEXT` everywhere.

Thirdly, use case. While it may seem neat to have MySQL support, who's
actually going to use it? For something like Condict, there are very few
real advantages. For desktop use, there will never be more than one
reader and one writer at any given time. For use on someone's web
server, it is exceedingly unlikely that there would be enough readers or
writers for SQLite to suffer even slightly.

As I see it, there are no real reasons to keep MySQL support around.
Focusing on SQLite simplifies development and maintenance.

In the future, we can simplify the code additionally, e.g. removing
unnecessary things from the database schema definitions.