Remove package @condict/gen-id
Rather than try to solve this problem ourselves, it's better to use a
more established solution - in this case, nanoid. Less code to maintain,
and better randomness and customisability to boot.

This commit also adds two new functions to `@condict/ui`:

- `genUniqueId()`, which generates a unique ID suitable for HTML
  elements
- `useUniqueId()`, a hook version of the above.

I also snuck in a little `user-select: none` for menus.