server: Refactor models and mutators to be "plain" objects
Instead of the previous solution, where we constructed a pair of magical
objects with model and mutator resolvers in a messy and slightly
convoluted fashion, all models and mutators are now plain objects with
plain functions. The only purpose of the automagical model and mutator
instantiation was to provide the current database connection, which we
can just as easily pass as a regular argument.

This new organisation also makes it much easier to track dependencies,
as we `import` models and mutators like any other JS value. No need to
invent a poor-man's dependency injector.