ui: Make commands more type safe
The `exec` property of a command is now generic, and doesn't even have
to be a function. As a result, `onExec` is mandatory, as we can't make
any assumptions about how to execute the command.

Note that `BoundCommand` looks the same. It effectively hides the
generic `exec` type by delegating to `onExec`.