server: Remove import/export functionality
This functionality was mainly intended for interop between MySQL and
SQLite, when moving from one to the other. However, the only database
engine now is SQLite, so it makes much less sense to keep this code
around, especially since SQLite stores everything in a single file.

Additionally, the actual implementation, while functional, had a variety
of edge cases that were not handled correctly (e.g. importing tags that
already exist in the target dictionary), and was unable to deal with
older schema versions at all.

Better to have no import/export at all, than to have one that is broken.

A similar functionality *may* be implemented in the future, for merging
existing dictionaries, but not for a while.