rte: Save previous non-null selection on editor and use it
When the native DOM selection changes to something outside the Slate
editor, Slate sets the editor's selection to null. This means, among
other things, that toolbar items don't work. Moreover, tabbing back to
the editor resets the selection to the start. By saving the previous
non-null selection, we can provide a better experience for the user.

This commit also updates various code to make use of the new value, if
the editor doesn't have a valid selection.