ui: Replace makeColorFn with a less awful solution
The previous `makeColorFn` thing was not particularly flexible. In
practice it's actually not that useful to be able to specify only TWO
levels of saturation (it's one reason Condict's colours are not great).
Raw `hsl` values are much more flexible, even if they require slightly
more maintenance. In the interest of reducing duplication, we now
export various hue constants instead of weird colour functions.