ui: Rewrite DescendantCollection to new Descendants type
The previous `DescendantCollection` was a bit of a mess. It put too
many requirements on its descendants, in particular in demanding they
all on some level map to an HTML element. The new `Descendants` only
requires that the descendants can be ordered in some fashion, so that
the `first()`, `last()`, `next()`, `prev()`, etc functions can be
implemented in a reasonable fashion. It's also a "plain old object"
without class prototype: I favour this style largely because each helper
function can then impose additional constraints on the item type, if
necessary. It feels cleaner.

There is also a new hook, `useDescendant()`, which makes it slightly
more ergonomic to register an item in a descendant collection.