ui: Refactor commands to be more composable
The problem with the previous solution was `<CommandGroup>`. It did
three distinct things:

1. It defined and initialised the command mapping;
2. It exposed commands in the group to child components; and
3. It rendered an element that listened to key events.

The new implementation decouples these two concerns. The first concern
is taken care of by the new hook `useCommandGroup`. The second concern
is handled by the new `<CommandProvider>`. The third and final concern
is required to be handled by the creator of the command group. Commands
are in other words uninvasive; they no longer render an element that
listens to key events. This makes commands more composable and easier to
fit into other components without introducing unnecessary wrapper divs
or fiddling with `as` props.