server: Make resolvers more type-safe
The new `FieldArgs` type generated by the GraphQL typer means we can now
create `IFieldResolver` values with exact argument types, nice!

Also, the `ResolversFor` type can figure out from type information which
fields require resolvers. The default behaviour in Apollo is to resolve
fields to a matching property on the 'parent' object, so naturally
resolvers are required for fields that don't have a matching property.

Lastly, using `null` as the root value makes the typings a bit easier to
work with. Notably, the `P extends null` checks in `RequiredResolvers`
and `OptionalResolvers` would not work with unknown: every type extends
`unknown`.