Adjust paddings, margins, sizes and other things, oh my
The basic idea is to try to fit UI components and text into an 8x8 grid
as far as it is reasonable to do so. It makes the components look a bit
more aligned and regular. This commit also reduces the font size
somewhat, bringing it more in line with other desktop apps, and makes
checkboxes, radio buttons and switches smaller.

Some additional component demos are included so that alignment of icons
can be tested too.