'~' is no longer a special stem name
There is no particularly good reason to treat `{~}` specially as a stem.
While it may seem nice to have a stem specifically for the lemma form,
this design decision also means it's harder to design inflection tables
if you have to constantly remember that `{~}`, for totally arbitrary
reasons, cannot be configured with a different value later on.