Make mutation results nullable
When mutation results are non-nullable, any failure that occurs will
effectively erase the results of other mutations in the query. That is,
given an operation like this:

  mutation {
    foo(data: "Foo") { ... }
    bar(data: "bar") { ... }
  }

if the mutation `foo` fails and does not have a nullable return type,
then the result of `bar` will be inaccessible, because the returned JSON
will have

  {
    "data": null,
    "errors": [...]
  }

With nullable mutation result types, a failing mutation only causes its
own result to become null.

This change was made to all mutations, including those that already have
ways of signalling failure (e.g. `logIn`), to safeguard against
unexpected errors.