server: Update to latest better-sqlite3
With this update, SQLite has support for the `RETURNING` clause. This
means we can write updates and inserts that return data, which might be
useful in future. It also means we cannot rely on `stmt.reader` to
determine whether a statement mutates the database or not, so the
Accessor class has been updated to take this into account.

Additionally, `@types/better-sqlite3` is unfortunately woefully out of
date (and, in my opinion, not terribly well-written), so I wrote my own
type definitions.