server: Implement basic search functionality
This requires several new tables for full-text search indexes, methods
for mutating them, and a variety of other things. The exact capabilities
of the search engine are informed largely by the features of SQLite, and
the search feature is still pretty rudimentary.

The choice of token characters for the FTS tables is largely arbitrary,
and is intended to capture a generic cross-linguistic notion of "word".
It does include `'` but does not include `-` or other separator
punctuation. Diacritics are *not* stripped, which is almost guaranteed
to be wrong. The alternative is also almost guaranteed to be wrong.

There is currently no way to specify which part of a description should
be searchable; the entire text will be indexed. This may be fixed in the
future.