app: Add first implementation of in-tab side panels
These panels will be used for editing of data and maybe auxiliary
information. They are modal with respect to the open tab: when a tab has
a side panel, the tab's regular content is inaccessible. Panels can open
subpanels, in which case only the deepest panel can be interacted with.

When a tab has an open panel, other tabs can still be accessed, and it's
still possible to show dialogs, open the search, navigate around, and so
on. Panels can mark themselves as having unsaved work, in which case the
owner tab will also be marked dirty.

When a panel is open, it is not possible to navigate within the tab
page.