app: Allow sidebar tabs to be focused by mouse (sometimes)
If focus is already inside the sidebar, then mousedown on a tab moves
focus to the tab, instead of keeping it on whatever was focused in the
sidebar before. It felt really weird to have focus on the search button,
then click a tab and have the search button remain focused. The basic
idea now is that if the sidebar "owns" the focus, then focus can move
around within the sidebar; and if the main area "owns" the focus (or the
focus is unowned), then the main area will claim it.