server: Fix more logic around FTS query formatting
For a query like `foo "`, the quoted phrase group of the regex will
match (since there is a double quote), but `m[1]` will be the empty
string. Instead of checking `m[1]`, we test that `m[3]` (the non-quoted
phrase group) is not null or undefined. Only one or the other can match.