app: Add hot reloading of translations
This change makes it much easier to edit translations, as you no longer
have to restart the app for them to take effect. It even detects the
addition and removal of locales.

The one caveat is that tab titles are *not* affected by the hot reload,
as they are fixed when the tab is created. Nevertheless, now you'll
never need more than a window reload to get fresh translations.