Expose a WritingDirection for programmatic handling of writing direction
In right-to-left languages, left and right arrows generally need to be
swapped. By exposing the current writing direction as a context, we can
make use of it in components for dynamic behaviour.

The ZoomSlider in the app his updated to take advantage of this new
context. Other components will follow.