ui: Add support for borderless buttons
These light up the entire button when hovered or pressed. Normally the
border is not highlighted. This makes it much easier to place the button
on backgrounds with the same colour as the button, where you want the
entire button to light up.

I also cleaned up the button styles a touch; there was far too much
repetition. And I made sure to use borderless styles where appropriate
(except in the app; that'll be a small separate change).