server: Add timeCreated and timeUpdated to various resources
In a future commit, it will be possible to query recently created and
recently updated things. The idea here is to give the app a bit of
liveness by showing what's been done recently. As a pleasant extra
effect, timestamps make it easier to find your way back to things you've
been working on lately.

Additionally, this commit renames the GraphQL `Date` type to
`UtcInstant`, to better capture what it actually represents.