server: Add query plan logging
This is used to debug query performance by providing a high-level
overview of what the query actually does. Chiefly, this is used to
ensure SQLite chooses appropriate indexes.

Query plan logging can be enabled by setting the environment variable
`DEBUG_QUERIES` to `plan`. The associated query is always logged before
the plan; otherwise it would be impossible to see which query the plan
applies to.