app: Implement (mostly) functional first version of home screen
This commit got a lot bigger than I would have liked. In addition to
implementing a (mostly) functional first version of the app's home
screen, it contains numerous tweaks and a few minor refactors:

- Main screen components are moved from the app's `ui` folder to
  `screens/main`, to keep them closer to the main screen and avoid
  circular dependencies.
- The `Page` type is moved out of `pages` and into its own file, also to
  prevent circular dependencies.
- A new `TagPage` type is added so links to tags can have somewhere to
  lead.
- Tag icon changed from a label with text to '#'. I'm not 100% happy
  with this, but people associate '#' with generic metadata so strongly
  that it seems silly not to take advantage of pre-ingrained knowledge.
- Added a rich-text renderer that can format descriptions and other rich
  content as returned by GraphQL. We currently use this to render
  language descriptions, but it will of course be used for definitions,
  table captions and possibly more.
- Added a new `DataViewer` component that can map data context results
  to loading, error and data screens. This should simplify a variety of
  data rendering. I wanted to call it `DataView`, but that's taken by a
  builtin.
- Font sizes and line heights are tweaked for pretty much all heading
  sizes (and will probably be tweaked again and again).
- Font sizes and line heights are normalised across more of the code
  base. Perhaps these should be moved to the theme.

The "Add language" button on the home screen is *not* functional yet. I
originally wanted to make it part of this commit, but additional changes
are needed before I can implement that functionality.

I also intend to add a new `statistics` field to the `Language` GraphQL
type, so we don't have to fetch `partsOfSpeech` and `tags` just to get
counters.

Lastly, aw yes commit 666.