app: Flesh out language page
This includes a proper header with a (functioning) Edit button, an About
section with the language description, and a WIP part of speech list. A
few components have been restructured to make room for these changes.

Additionally, the language form now marks the tab as dirty if you've
made changes. Nice!