app: Update ConfirmButton to confirm on release
If the confirm action is triggered during the keydown event, the key may
be transferred to whatever gets focus once the panel closes, which can
result in strange behaviour. Confirming on release means the user has to
abort the key press first, which seems to work much more reliably.

This change introduces a new prompt to let the user know they have to
release as well. And the timeout is reduced to 2 seconds to account for
time spent reading the prompt.