table-editor: Use position: fixed for popup
The table editor needs to be able to appear inside containers with
scrollable overflow. The previous implementation did not permit that; it
would have caused the popup to overflow too. Now it acts more like a
"true" popup.

Performance-wise, updating the position in an animation frame is
probably not ideal, but the editor is unlikely to open for a long time
anyway. Plus, I've tried to limit DOM updates by only writing to `style`
when things change.