app: Reword 'derive lemma'-related options
The phrasing "not added to dictionary" is extremely strange and
confusing. Surely the whole point of creating an inflection table is to
add it to the dictionary, right? What does it mean for a form not to be
added?

Hopefully the phrasing "Create headword from this form" will be less
confusing. The feature is still kind of subtle and will probably require
some reading for most users to make sense of, but for more experienced
users, it's best to avoid a confusing choice of words.

Also squeezed in another small l10n change.