More visible focus indicators!
Instead of a blurry shadow, let's go crisp and clear. It's more visually
distinct, and less visually noisy. Because it's so clear, it's also
easier to rest your eyes on, and feels less like a glowy distraction
(especially on a dark background).

Additionally, buttons get a subtle inset shadow that helps to delimit
the focus indicator from the button itself. This helps especially when
there is strong contrast between the button and the background it's on
(so mainly bold buttons).