app: Abstract resource times into a shared ResourceTime
There is a little bit of logic around deciding whether to show
timeCreated or timeUpdated, and this component simplifies the process.
Moreover, in a future commit, it will be possible to sort recent changes
by creation or update time, and being trivially able to switch to
creation time is useful.