app: Use custom validation instead of HTML5 validation
HTML5 validation in forms unfortunately doesn't really allow you to
intercept validation on submit when used with `react-hook-form`. For
attributes like `required`, that means the user gets a popup that says
"Please fill in this field" or similar, and the app is given no
opportunity to show its custom error message.

Moreover, the HTML5 validation error popup isn't really customisable.
For l10n, this means I can't easily translate the message.

And lastly, the browser only shows the first error. If you have multiple
errors, you'll be taken to each one in turn as you fix the errors.
Having an overview of all errors is *probably* perferable. Probably.

Of course, users still need to know what's expected, so `aria-required`
is put on fields that are required, and `aria-invalid` on fields that
have errors.

I decided to sneak in a `SelectField` component that will be necessary
for the upcoming definition form.

Lastly, I changed the validation mode from `onTouched` to `onSubmit`,
mainly to prevent things from jumping around as you tab through the
form. It's really quite jarring; I see why `onSubmit` is the default.