table-editor: Disable focus trap before committing cell
When the cell value is committed, the cell dialog is immediately
destroyed. As a result, there is no element that the browser to emit a
`blur` event on, so it simply doesn't happen. If the user clicks outside
the table, we never become notified that the table has lost focus, and
the table behaves strangely as a result.