ui: Add `openClass` to `MenuTrigger` and `ContextMenuTrigger`
This new class makes it much easier to inject simple custom styling when
the menu is open, and also isolates renders to the tree inside the
trigger. For more advanced uses, `onToggle` remains available.

The `Button` and `LinkButton` components have been updated with styles
for `.force-active` primarily to help support this new prop (and
simplify some styles in the app).

`Toolbar.MenuButton` has been updated to take advantage of this new
feature.