app: Overcomplicate definition inflection table handling
react-hook-form has a number of quirks surrounding `useFieldArray`, and
several around inheritance of default values too. This new solution is
hacky, messy, convoluted and undesirable, but fixes one important bug:

1. Edit a definition
2. Remove a existing inflection table
3. Add a new inflection table

The new table would have the layout, caption and other values of the
form's defaultValues for that particular table index. This is confusing
and impossible to work around in a satisfactory manner.