Z Trek II: The Wrath of Index
z-index is dumb and it should feel dumb.

Anyway, as part of this commit, I think I will endeavour to use
`z-index: 10` when I need something (like a dialog or popup) to be
"definitely above" everything else, and values 1–9 can be used for
inbetweeny things. As an additional, special guarantee, 6–9 are always
above the form buttons at the bottom (they're the biggest source of pain
right now).

I'm sure this will work great until I forget it and cause more problems
in a few commits.