ui: Remove useless layout components
The `Panel`, `Grid` and `Cell` components were simultaneously too
generic and too specific. I envisioned myself using them in the app for
basic layout things, but that ended up not happening. Grids aren't
actually that difficult to define in CSS, and the panel thing was not
something I actually needed or wanted.

Hence: goodbye, components!