app: Break out debounced persisting and rework slightly
Abstracting this feature into a separate thing is intended to make it
easier to add support for sessions and other things in a future commit,
in addition to cleaning up the code somewhat. The new implementation
requires a logger, so we now create the logger in the config loader
instead of in the main function. As a bonus, this allows us to filter
out those "saved app config" messages that previously cluttered the
console whenever you changed the config.